move ref() and Ref.fromUrl() to utils
