+++ packages/refs/src/utils.ts
@@ -1,8 +1,31 @@
 import { DocHandle, Repo } from "@automerge/automerge-repo";
-import type { MatchPattern, CursorMarker, AutomergeRefUrl } from "./types";
+import type {
+  MatchPattern,
+  CursorMarker,
+  AutomergeRefUrl,
+  PathInput,
+  AnyPathInput,
+} from "./types";
 import { CURSOR_MARKER } from "./types";
 import { Ref } from "./ref";
 import { parseAutomergeRefUrl } from "./parser";
+
+/**
+ * Create a ref with automatic type inference.
+ *
+ * @example
+ * ```ts
+ * const titleRef = ref(handle, 'todos', 0, 'title');
+ * titleRef.value(); // string | undefined
+ * ```
+ */
+
+export function ref<TDoc, TPath extends readonly PathInput[]>(
+  docHandle: DocHandle<TDoc>,
+  ...segments: [...TPath]
+): Ref<TDoc, TPath> {
+  return new Ref<TDoc, TPath>(docHandle, segments as [...TPath]);
+}
 
 /**
  * Create a cursor-based range segment for stable text selection.
@@ -17,6 +40,24 @@ import { parseAutomergeRefUrl } from "./parser";
  */
 export function cursor(start: number, end: number): CursorMarker {
   return { [CURSOR_MARKER]: true, start, end };
+}
+
+/**
+ * Parse a ref from an Automerge URL string.
+ *
+ * @param handle - The document handle to use
+ * @param url - Full automerge URL like "automerge:documentId/path#heads"
+ *
+ * @example
+ * fromUrl(handle, "automerge:abc/todos/0#head1|head2" as AutomergeRefUrl)
+ */
+export function fromUrl<TDoc = any>(
+  handle: DocHandle<TDoc>,
+  url: AutomergeRefUrl
+): Ref<TDoc, AnyPathInput[]> {
+  const { segments, heads } = parseAutomergeRefUrl(url);
+  const options = heads ? { heads } : {};
+  return new Ref<TDoc, AnyPathInput[]>(handle, segments, options);
 }
 
 /**
@@ -37,7 +78,7 @@ export async function findRef<T = any>(
   const handle = await repo.find(documentId as any);
   await handle.whenReady();
 
-  return Ref.fromUrl(handle as DocHandle<T>, url);
+  return fromUrl(handle as DocHandle<T>, url);
 }
 
 /**
